requirements.txt
accelerate==0.26.0
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==24.2
Pillow==11.0.0
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.49.0
triton==3.0.0
typing_extensions==4.9.0
urllib3==2.3.0